+++ test_settings.py
@@ -1,7 +1,11 @@
+USE_TZ = True
+
+TIME_ZONE = 'Africa/Johannesburg'
+
 DATABASES = {
     'default': {
-        'ENGINE': 'django.contrib.gis.db.backends.spatialite',
-        'NAME': 'jmbo.db',
+        'ENGINE': 'django.contrib.gis.db.backends.postgis',
+        'NAME': 'jmbo',
         'USER': '',
         'PASSWORD': '',
         'HOST': '',
@@ -21,6 +25,14 @@ INSTALLED_APPS = [
     'photologue',
     'secretballot',
     'publisher',
+    'south',
 ]
 
 SITE_ID = 1
+
+# Disable celery
+CELERY_ALWAYS_EAGER = True
+BROKER_BACKEND = 'memory'
+
+# xxx: get tests to pass with migrations
+SOUTH_TESTS_MIGRATE = False